64-gce-disk-removal.rules: fix error-handling
This rule tries to unmount a block device after it has been detached.

The `umount && logger` pattern ensures we only log a warning if umount
succeeded. However, if it did not, the overall exit code of the shell
command is still nonzero, which systemd will log as a udev error like:

    (udev-worker)[534]: vda2: Process '/bin/sh -c '/bin/umount -fl
    /dev/vda2 && /usr/bin/logger -p daemon.warn -s WARNING: hot-removed
    /dev/vda2 that was still mounted, data may have been corrupted''
    failed with exit code 32.

Fix this by converting the logic into an if-statement. (We could
alternatively use `! umount || logger` though that's less obvious.)